Move `Decompressor` into `decompressor.rs`
